# Refactor pushing and validation into their own functions