@MasslessParticle 2022-09-12
Benchmarks
old:
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/distributor
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
Benchmark_Push-8                     	      26	  43777618 ns/op	 6405701 B/op	  200075 allocs/op
Benchmark_PushWithLineTruncation-8   	      30	  43495762 ns/op	 6405626 B/op	  200076 allocs/op
PASS
ok  	github.com/grafana/loki/pkg/distributor	3.917s



New:
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/distributor
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
Benchmark_Push-8                     	      27	  41896587 ns/op	 6405636 B/op	  200074 allocs/op
Benchmark_PushWithLineTruncation-8   	      27	  42563465 ns/op	 6405481 B/op	  200075 allocs/op
PASS
ok  	github.com/grafana/loki/pkg/distributor	3.882s



cmp:
benchmark                              old ns/op     new ns/op     delta
Benchmark_Push-8                       43777618      41896587      -4.30%
Benchmark_PushWithLineTruncation-8     43495762      42563465      -2.14%

benchmark                              old allocs     new allocs     delta
Benchmark_Push-8                       200075         200074         -0.00%
Benchmark_PushWithLineTruncation-8     200076         200075         -0.00%

benchmark                              old bytes     new bytes     delta
Benchmark_Push-8                       6405701       6405636       -0.00%
Benchmark_PushWithLineTruncation-8     6405626       6405481       -0.00%