@MasslessParticle 2022-09-13
Closing in favor of a design doc detailing a different approach